Automated Deployment (#2)
Only files in the root directory of a branch or docs
directory can be hosted on GH-Pages. Since Markbind builds its
source code to the `_site` directory, Github will not automatically
deploy the updated site.

Let's add a workflow that will build the updated site and push it to
the `gh-pages` branch's root directory, thus automatically updating
the GH Pages site.