Write documentation on creating custom CATcher sessions (#19)
Write documentation on creating custom CATcher sessions

The instructions on loading custom sessions in the Testing page of the
DG does not specify how to create these sessions. This causes new
developers to get confused about the purpose of this section.

Let's add some documentation to explain how to create these custom
sessions, and also explain how to populate the repositories with issues.